added task execution configuration option: shell
a task's `shell` setting may be a string or a mapping of:

  {
    script: str,
    executable: str
  }

as a simple string, `shell` will be executed by "sh".

supported executables include: sh, bash and python.

all executables are launched passing the shell "script" string as an
argument to "-c", *e.g.*:

  sh -c "..."

the shell "script" supports template expression interpolation (like
"exec") -- in addition to, of course, shell-specific variables and
syntax.

for example:

  clean-tmp:
    shell: find /tmp/ -mindepth 1 -type d -empty -delete

  go-nuts:
    shell:
      executable: python
      script: |
        import subprocess

	proc = subprocess.run(['cowsay', '-e', '{{ env.DISPLAY }}'])

	raise SystemExit(proc.returncode)

note: this is really just syntactic sugar for the following and can be
custom-implemented like the following:

  exec:
    - perl
    - -e
    - |
      print "Hello World";

(or, obviously, via a script file on disk).