task state persistence
initial working draft of task state persistence via anonymous pipe
communication with the task subprocess.

each task subprocess is provisioned with a state-in pipe and a state-out
pipe -- dup'd to file descriptors 3 and 4 (respectively) -- for simple
and conventional readable input of a task's existing state and writable
output of a task's updated state for persistence.

the `fate.task` sdk package is extended with helpers for reading and
writing (structured) state in sub-package `fate.task.state`.

task state is persisted on disk via lmdb.