prevent error due to multiple invocations of ScheduledTask.ready() / .poll() (during introspection)
while this does not occur during current operation, it was exposed
during debugger introspection, that multiple invocations of this method
on a completed task raises an error (due to an attempt to read the task
state pipe multiple times).

this fix prevents the error.

(note that future plans to refactor this class may further prevent such
a state of affairs.)