conf: if/unless: permit boolean type (in addition to str template) expressions
the task configuration key `if` (and its convenience complement
`unless`) are *generally* presumed to be sophisticated expressions -- of
configuration format type string -- which are rendered as templates in
order to evaluate their boolean result.

however, it is reasonable for a user to set `if: false`, for example, in
order to statically disable a task (temporarily) (without removing its
configuration all together). such values will be parsed as booleans --
not strings -- under YAML and TOML.

and, we may as well support this -- rather than forcing these to be
_e.g._ `if: 'false'` -- and lighten the load on the value's
interpretation by task configuration property `if_`.