sched: disallow zero-value interval
the tiered-tenancy scheduler currently utilizes short sleeps between
cycles of polling its active tasks. previously, the length of these
waits was derived from the OS scheduler's round-robin quantum, if
available to query.

however, this value may appropriately be ZERO. while this is "fair," and
not catastrophic, it does undo the utility of the wait.

while we _could_ leave this as is, for the time being we'll handle this
situation with the *default* interval, such that we do indeed release
the CPU for some length of time, before polling again.

(See also the discussion of this issue in testing, where it was
uncovered: 84586650516193940db1dceb406b56d01c511aae.)