core: strip out plumbum from subprocess execution -- and thereby redress task output infidelity
* plumbum was treating *all* subprocess outputs as utf-8 -- and dropping
  characters which would not decode this way -- and thereby corrupting
  binary (and binary-adjacent) outputs

* future work will only cause plumbum to get more in the way (result
  archives, log-streaming, etc.)

though plumbum is useful in other contexts, and though we get plumbum
"for free" with argcmdr, (and which *is* useful for our CLI definition),
plumbum doesn't really do much of anything for us here -- we were
essentially using it as a very thin and superfluous layer over
subprocess.

in removing plumbum, we're forced to handle encoding/decoding -- but
this is something we have to do -- (and something which subprocess could
also do for us, but which we have to do ourselves).